Pongo el div de loading y errores

+++ src/js/app/core/comment/comment.service.js
@@ -5,32 +5,62 @@ require('angular-local-storage');
 
 angular.
   module('core.comment').
-  factory('Comment', ['localStorageService',
-    function(localStorageService) {
-
+  factory('Comment', ['localStorageService', '$q',
+    function(localStorageService, $q) {
       return {
         query: function(symbolId) {
-          var commentKey = 'comment-' + symbolId + '-';
-          return localStorageService.keys()
-                  .filter(function(key) {
-                    return key.startsWith(commentKey);
-                  }).
-                  map(function(item) {
-                    return localStorageService.get(item);
-                  }).
-                  sort(function(a, b) {
-                    return new Date(b.date) - new Date(a.date);
-                  });
+          // Someday this function wil be async
+          var defered = $q.defer();
+          var promise = defered.promise;
+          setTimeout(function() {
+            if (localStorageService.isSupported) {
+              var commentKey = 'comment-' + symbolId + '-';
+              var comments = localStorageService.keys()
+                      .filter(function(key) {
+                        return key.startsWith(commentKey);
+                      })
+                      .map(function(item) {
+                        return localStorageService.get(item);
+                      })
+                      .sort(function(a, b) {
+                        return new Date(b.date) - new Date(a.date);
+                      }); 
+              defered.resolve(comments);
+            } else {
+              defered.reject({"browser": "Local Storage not supported"});
+            }
+          }, 500);
+          return promise;
         },
         save: function(symbolId, commentId, comment) {
-          var key = 'comment-' + symbolId + '-' + commentId;
-          localStorageService.set(key, comment);
-          return this.query(symbolId);
+          // Someday this function wil be async
+          var defered = $q.defer();
+          var promise = defered.promise;
+          setTimeout(function() {
+            if (localStorageService.isSupported) {
+              var key = 'comment-' + symbolId + '-' + commentId;
+              localStorageService.set(key, comment);
+              defered.resolve(comment);
+            } else {
+              defered.reject({"browser": "Local Storage not supported"});
+            }
+          }, 500);
+          return promise;
         },
         remove: function(symbolId, commentId) {
-          var key = 'comment-' + symbolId + '-' + commentId;
-          localStorageService.remove(key);
-          return this.query(symbolId);
+          // Someday this function wil be async
+          var defered = $q.defer();
+          var promise = defered.promise;
+          setTimeout(function() {
+            if (localStorageService.isSupported) {
+              var key = 'comment-' + symbolId + '-' + commentId;
+              localStorageService.remove(key);
+              defered.resolve(commentId);
+            } else {
+              defered.reject({"browser": "Local Storage not supported"});
+            }
+          }, 500);
+          return promise;
         },
       };
     }
